+++ internal/auth0/connection/okta/resource_test.go
@@ -1,4 +1,4 @@
-package connection_test
+package okta_test
 
 import (
 	"fmt"
